+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Setup local-only integration tests
 - Fix: Level should be positive
 - `PUT /game`: Return `400` if timestamp is older than the last game saved
+- Expose `playerIdentifier` in `/high-scores` composed of `external_id` and `sso_platform`
+- Refactor `scoreService.ts`
+- In `players` table, rename `platform` field to `sso_platform`
 
 ## 0.1.6 - 08/11/2025 18:35 UTC+1